requirements.txt
accelerate==1.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
attrs==24.2.0
blinker==1.9.0
Byaldi==0.0.7
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
colpali_engine==0.3.4
datasets==3.1.0
dill==0.3.8
eval_type_backport==0.2.0
exceptiongroup==1.2.2
faiss-cpu==1.9.0.post1
filelock==3.13.1
frozenlist==1.5.0
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
GPUtil==1.4.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.5
idna==3.10
Jinja2==3.1.3
joblib==1.4.2
jsonlines==4.0.0
jsonpath-python==1.0.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mistralai==1.2.5
ml_dtypes==0.5.0
mpmath==1.3.0
mteb==1.6.35
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.16.0
networkx==3.2.1
ninja==1.11.1.2
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
peft==0.11.1
pillow==10.4.0
propcache==0.2.1
protobuf==5.29.1
psutil==6.1.0
pyarrow==18.1.0
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytrec-eval-terrier==0.5.6
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.1
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
srsly==2.4.8
streamlit==1.40.2
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.20.3
toml==0.10.2
torch==2.5.1+cu124
tornado==6.4.2
tqdm==4.67.1
transformers==4.46.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
xxhash==3.5.0
yarl==1.18.3
